Allow subject viewer to determine rendering component
