+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/SingleImageViewer/index.js
@@ -1,1 +1,1 @@
-export { default } from './SingleImageViewer'
+export { default } from './SingleImageViewerContainer'
